- 修改心跳报文到独立的线程接受处理。
