+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - 修改部署设计子表的表结构，增加连接父表的外键。
 - 设备表中添加操作系统平台字段，实现部分加入相应的处理逻辑。
 - 修改设备表和组件表建立和工程表之间的外键关联。
+- 修改心跳报文到独立的线程接受处理。
 ### Fixed
 - 修复了部署设计父子表、组件父子表的及联保存的问题。
 - 修复多个网卡的情况下发送一个网卡ip的问题。
