+++ src/main/java/com/rengu/operationsoanagementsuite/OperationsoanagementsuiteApplication.java
@@ -1,6 +1,6 @@
 package com.rengu.operationsoanagementsuite;
 
-import com.rengu.operationsoanagementsuite.Thread.UDPReceiveThread;
+import com.rengu.operationsoanagementsuite.Thread.UDPHearBeatReceiveThread;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -12,6 +12,6 @@ public class OperationsoanagementsuiteApplication {
     public static void main(String[] args) {
         SpringApplication.run(OperationsoanagementsuiteApplication.class, args);
         // 启动UDP监听线程
-        new UDPReceiveThread().run();
+        new UDPHearBeatReceiveThread().run();
     }
 }
